#25 Termino las funcionalidades de Agregar,Actualizar y Eliminar de los Tipos de Servicios y Servicios. Agrego archivo para ordenar los endpoints en swagger. Agrego descripcion de las apis y endpoints para la documentacion en swagger. Agrego permisos a los endpoints creados

+++ src/main/java/com/cperalta/jardineria/servicios/infraestructure/entities/ServicioEntity.java
@@ -30,7 +30,7 @@ public class ServicioEntity {
     @Column(unique = true)
     private String descripcion;
 
-    private Long precio;
+    private Float precio;
 
     @ManyToOne(cascade = CascadeType.MERGE)
     @JoinColumn(name = "tipodeservicio_id", referencedColumnName = "id")
